+++ src/components/Header/Header.css
@@ -1,0 +1,77 @@
+body {
+  margin: 0%;
+}
+button {
+  margin: 30px auto;
+  padding: 8px;
+  border: none;
+  box-sizing: border-box;
+  width: 50%;
+  cursor: pointer;
+  display: inline-block;
+  text-align: center;
+  border: 3px solid #5f5f5f;
+  border-radius: 5px;
+  background-color: #5f5f5f;
+  color: #fff;
+  transition: 0.3s all;
+}
+button:hover {
+  border: 3px solid #5f5f5f;
+  background-color: #efefef;
+  color: #2f2f2f;
+}
+.backgroundImage {
+  width: 100%;
+  height: 100%;
+}
+.onImage {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0%;
+  left: 0%;
+}
+.logo {
+  margin: 10px;
+  font-size: 30px;
+}
+.backgroundForm {
+  background-color: rgba(255, 255, 255, 0.5);
+  border-radius: 20px;
+  margin: 0px auto;
+  padding-top: 10px;
+  padding-bottom: 30px;
+  width: 60%;
+}
+.pageTitle {
+  text-align: center;
+}
+.formItem {
+  margin-top: 20px;
+  margin-left: 25%;
+}
+.formInput {
+  margin-left: 10%;
+  width: 50%;
+}
+.buttonCenter {
+  text-align: center;
+}
+
+.headerBack {
+  background-color: rgb(255, 255, 255);
+  width: 100%;
+  height: 20%;
+  margin: 0px auto;
+}
+.headerLogo {
+  display: inline-block;
+  margin: 10px;
+  font-size: 30px;
+}
+.headerButton {
+  display: inline-block;
+  margin: 0 0 0 auto;
+  text-align: center;
+}
